[Feat] Database 설계 및 Icon 연동
* Firebase에 저장할 Database 설계
* Firebase <-> Local Icon 연동
