+++ Assets/05_KGW_Folder/Scripts/UI/DefeatChapterUI/DefeatChapterUI.cs
@@ -28,8 +28,6 @@ public class DefeatChapterUI : BaseUI
     // 로비 이동 버튼 클릭
     private void LobbyButtonClick()
     {
-        // TODO : 김기운 : 추후에 마이씬 매니저 교체 예정
-        // GameManager.Instance.Scene.LoadSceneAsync(SceneName.SDW_LobbyScene, UIName.KGW_StageSelectUI);
         GameManager.Instance.Scene.LoadSceneAsync(SceneName.SDW_LobbyScene);
         OnUICloseRequested?.Invoke(UIName.DefeatChapterUI);
     }
