+++ Assets/06_SDW_Folder/Scripts/Manager/UIManager.Open.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eb51d63d7c0e43468b1c365d65419950
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
