+++ Assets/06_SDW_Folder/Scripts/Manager/UIManager.cs.meta
@@ -1,11 +1,3 @@
-fileFormatVersion: 2
-guid: eb51d63d7c0e43468b1c365d65419950
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
+﻿fileFormatVersion: 2
+guid: 94e358302f16403a81622e1d89f325dd
+timeCreated: 1757833549
