+++ Assets/06_SDW_Folder/Scripts/Struct/UserInfo.cs
@@ -1,0 +1,18 @@
+﻿namespace SDW
+{
+    public struct UserInfo
+    {
+        public string Nickname;
+        public string Email;
+        public string UserId;
+        public int IconNumber;
+
+        public UserInfo(string nickname, string email, string userId, int iconNumber)
+        {
+            Nickname = nickname;
+            Email = email;
+            UserId = userId;
+            IconNumber = iconNumber;
+        }
+    }
+}
