+++ Assets/06_SDW_Folder/Scripts/Struct/UserInfo.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: e2ee7ecd4dc244c993ccfa6d6023bf75
+timeCreated: 1757830221
